move dkg module to its own file; update and refactor gendoc
